feat: relocate form contact to about site and create new page for service
