+++ src/app/pages/service/page.tsx
@@ -1,0 +1,7 @@
+export default function Page() {
+  return (
+    <div className="mt-16 min-h-screen">
+      <h1>ajsdkaskld</h1>
+    </div>
+  );
+}
